+++ cart/domain/db/cart.go
@@ -1,4 +1,4 @@
-package domain
+package db
 
 import "go.mongodb.org/mongo-driver/bson/primitive"
