+++ cart/domain/repository/cartRepository.go
@@ -1,35 +1,38 @@
 package repository
 
 import (
-	"cartService/domain"
-	"cartService/internal"
+	"cartService/domain/db"
+	error "cartService/internal"
+
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-//type CartRepositoryDB interface {
-//	//Create(cart *domain.Cart) (*domain.Cart, *error.AppError)
-//	SaveItem(userId string, item domain.Item) (domain.Cart, error)
-//	RemoveItem(userId string, itemId string) (domain.Cart, error)
-//	UpdateItem(userId string, item domain.Item) (domain.Cart, error)
-//	GetItems(userId string) ([]domain.Item, error)
-//}
+type CartRepositoryDB interface {
+	Create(db.Cart) (*db.Cart, *error.AppError)
+	Read(id string) (*db.Cart, *error.AppError)
+	Update(db.Cart) (*db.Cart, *error.AppError)
+	Delete(db.Cart) (*db.Cart, *error.AppError)
+	DeleteAll() *error.AppError
+}
 
-type CartRepository struct{}
+type CartRepository struct {
+	// dbClient *mongo.Client
+}
 
-func (cdb CartRepository) Create(cart domain.Cart) (*domain.Cart, *error.AppError) {
+func (cdb CartRepository) Create(cart db.Cart) (*db.Cart, *error.AppError) {
 
-	newCart := domain.Cart{
+	newCart := db.Cart{
 		CustomerId: cart.CustomerId,
 		Products:   cart.Products,
 	}
 
 	newCart.Id = primitive.NewObjectID().Hex()
 
-	//ctx, cxl := context.WithTimeout(context.Background(), 10*time.Second)
-	//defer cxl()
+	// ctx, cxl := context.WithTimeout(context.Background(), 10*time.Second)
+	// defer cxl()
 
-	//orderCollection := Collection(cdb.dbClient, "cart")
-	//_, err := cartCollection.InsertOne(ctx, newCart)
+	// orderCollection := Collection(cdb.dbClient, "cart")
+	// _, err := cartCollection.InsertOne(ctx, newCart)
 
 	//if err != nil {
 	//	return nil, error.NewUnexpectedError("Unexpected error from DB")
@@ -38,4 +41,20 @@ func (cdb CartRepository) Create(cart domain.Cart) (*domain.Cart, *error.AppErro
 	cart.Id = newCart.Id
 
 	return &cart, nil
+}
+
+func (cdb CartRepository) Read(id string) (*db.Cart, *error.AppError) {
+
+	// ctx, cxl := context.WithTimeout(context.Background(), 10*time.Second)
+	// defer cxl()
+
+	// orderCollection := Collection(cdb.dbClient, "cart")
+	// order := db.Cart{}
+	// err := orderCollection.FindOne(ctx, bson.M{"_id": id}).Decode(&order)
+
+	// if err != nil {
+	// 	return nil, error.NewUnexpectedError("Unexpected error from DB")
+	// }
+
+	return &db.Cart{}, nil
 }
